# Optimize multiannotator.py for performance