--- a/tests/test_multiannotator.py
+++ b/tests/test_multiannotator.py
@@ -1,4 +1,3 @@
-import warnings
 from copy import deepcopy
 
 import numpy as np